refactor: 요구사항에 맞게 패키지 이동이 금지된 클래스 원상복구
